Fill HTML with content
All content has now been added to the file.

Initially, sidebar element icons (png) were added with a > img > "Home"
for example, but this was changed to svg once realised task was to use
svg and position via grid. However, left the a tags in since these would
normally require one on a real page.