Finish first draft of layout
Most items are in place and styled correctly at this point.

The main content area is styled with some responsive design, better
control of the trending area (required to move up) is needed. Further
enhancement would include making links more available on the cards.

For some reason when cards are stacked 3 card high, the bottom of the
trending box seems to be just a pixel or two higher than that of the
projects boxes, but i cannot visually fix it yet.

Some elements have been taken out of their container, others wrapped
inside new container for better styling.