Bring back list for sidebar
Decided it worked better with list on the sidebar. Also removed the
right column for the projects area as grid works well without it.

One thing I'm not able to do is to ensure the trending bit moves up and
down depending on where the cards are at and that the cards don't end
up with 4 and 2 cards. Since it seems like an easy task using
mediaquery I am leaving it for now until that is covered later.