Add media query
Added media query as I really did not like the 4/2 row set up on very
wide screen. As far as responsiveness, the page is very limited, but
the basics for the tasks are covered and with more knowledge it can
be enhanced later.

The right column does not want to align the way I would like it to,
that is to be addressed in the future reworks.